move byte/short to int

+++ desktop/applis/apps/adv/pokemonmock/src/test/java/code/mock/MockGamePkStreamTest.java
@@ -9,8 +9,8 @@ public final class MockGamePkStreamTest extends EquallableMockPkUtil {
     public void g1() {
         MockGamePkStream m_ = new MockGamePkStream();
         Game g_ = Instances.newGame();
-        g_.getDifficulty().setIvPlayer((short) 1);
-        g_.getDifficulty().setIvFoe((short)1);
+        g_.getDifficulty().setIvPlayer(1);
+        g_.getDifficulty().setIvFoe(1);
         m_.save("", g_);
         assertEq(1,m_.load("",null).getDifficulty().getIvPlayer());
         assertEq(1,m_.loadThen("",null).getDifficulty().getIvFoe());
