+++ desktop/applis/apps/bean/bean_games/pokemonbean/src/main/java/aiki/beans/fight/TeamBeanGetMembers.java
@@ -3,10 +3,9 @@ package aiki.beans.fight;
 import aiki.beans.PokemonBeanStruct;
 import aiki.beans.PokemonStandards;
 import code.bean.nat.*;
-import code.bean.nat.*;
 public class TeamBeanGetMembers implements NatCaller{
     @Override
     public NaSt re(NaSt _instance, NaSt[] _args){
-        return PokemonStandards.getByteArray(( (TeamBean) ((PokemonBeanStruct)_instance).getInstance()).getMembers());
+        return PokemonStandards.getIntArray(( (TeamBean) ((PokemonBeanStruct)_instance).getInstance()).getMembers());
     }
 }
